votechain/requirements.txt
flask==3.0.0
web3==6.11.3
py-solc-x==2.0.2
python-dotenv==1.0.0